+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 - **js** config: added a new option `allowedConsoleMethods` to control which `console` methods are allowed.
 - **jsdoc** config: added a new option `extraMultilineCommentsStartingWithToIgnore` to control which multiline comments should be ignored by [`no-bad-blocks`](https://github.com/gajus/eslint-plugin-jsdoc/blob/HEAD/docs/rules/no-bad-blocks.md) rule. Added `__PURE__` and `__NO_SIDE_EFFECTS__` to the default list.
 - **casePolice** config: autofix is now disabled by default.
+- **react** config: `configTypeAwareRules` now uses `files` and `ignores` from `ts` config by default.
 - Added a new root option, `defaultConfigsStatus`, to control what configs are enabled or disabled by default.
 
 ## 0.10.0